+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project are documented in this file, based on [Keep 
 
 
 ## [Unreleased]
+### Fixed
+- Always very verbose logging. Notes are now logged at Gradle INFO level, meaning that they are only shown with `--info` or `--debug`. Warnings are now logged at Gradle WARN level, so they can be ignored with `--warning-mode=none`. Errors are now logged at Gradle ERROR level, so they are always shown and outputted to stderr.
+- Language specifications rebuilding without changes due to `target/metaborg/table.bin` or `target/metaborg/table-completions.bin` being changed after the build.
+
+### Added
+- `spoofaxBuildApproximateAdditionalInputExcludePatterns` and `spoofaxBuildApproximateAdditionalOutputExcludePatterns` to `spoofaxLanguageSpecification` extension to support setting up additional include/exclude patterns when building language specifications.
 
 
 ## [0.5.3]